pythonProject/requirements3.txt
absl-py==0.12.0
astunparse==1.6.3
audioread==2.1.9
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==8.0.1
colorama==0.4.4
ConfigArgParse==1.4.1
cycler==0.10.0
Cython==0.29.23
dataclasses==0.8
DAWG==0.8.0
DAWG-Python==0.7.2
decorator==5.0.9
docopt==0.6.2
Flask==2.0.1
Flask-Cors==3.0.10
flatbuffers==1.12
gast==0.3.3
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==4.0.1
inflect==5.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jsonpickle==2.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
librosa==0.7.2
llvmlite==0.31.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==2.1.0
nltk==3.5
numba==0.48
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==8.2.0
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymorphy2==0.9.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pyworld>=0.2.10
PyYAML==5.4.1
regex==2021.4.4
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
rnnmorph==0.4.0
rsa==4.7.2
russian-tagsets==0.6
scikit-learn==0.24.2
scipy==1.5.4
six==1.15.0
SoundFile==0.10.3.post1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.8.1+cpu
torchaudio==0.8.1
torchvision==0.9.1+cpu
tqdm==4.60.0
typeguard==2.9.1
typing-extensions==3.7.4.3
Unidecode==1.2.0
urllib3==1.25.11
webrtcvad==2.0.10
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
pymorphy2-dicts-ru==2.4.417127.4579844
-e git+https://github.com/nsu-ai/russian_g2p#egg=russian_g2p
